Moved main glue application to glue.app
