+++ glue/qglue.py
@@ -200,10 +200,10 @@ def qglue(**kwargs):
         links = [(['balls.kg'], ['cones.lbs'], lb2kg, kg2lb)]
         qglue(balls=balls, cones=cones, links=links)
 
-    :returns: A :class:`~glue.qt.glue_application.GlueApplication` object
+    :returns: A :class:`~glue.app.glue_application.GlueApplication` object
     """
     from .core import DataCollection
-    from glue.qt.glue_application import GlueApplication
+    from glue.app.glue_application import GlueApplication
 
     links = kwargs.pop('links', None)
