+++ glue/qt/widgets/__init__.py
@@ -3,19 +3,7 @@ from .layer_tree_widget import LayerTreeWidget
 from .message_widget import MessageWidget
 from .mpl_widget import MplWidget
 
-from ...viewers.histogram.qt_widget import HistogramWidget
-from ...viewers.image.qt_widget import ImageWidget
-from ...viewers.scatter.qt_widget import ScatterWidget
-from ...viewers.dendro.qt_widget import DendroWidget
-
-default_widgets = [ScatterWidget, HistogramWidget, ImageWidget]
-
-try:
-    from ...viewers.ginga.qt_widget import GingaWidget
-    default_widgets.append(GingaWidget)
-except ImportError:
-    pass
-
 def enable_dendrograms():
-    if DendroWidget not in default_widgets:
-        default_widgets.append(DendroWidget)
+    from ..config import qt_client
+    from ...viewers.dendro.qt_widget import DendroWidget
+    qt_client.add(DendroWidget)
