+++ glue/qt/widgets/glue_mdi_area.py
@@ -19,7 +19,7 @@ class GlueMdiArea(QtGui.QMdiArea):
     def __init__(self, application, parent=None):
         """
         :param application: The Glue application to which this is attached
-        :type application: :class:`~glue.qt.glue_application.GlueApplication`
+        :type application: :class:`~glue.app.glue_application.GlueApplication`
         """
         super(GlueMdiArea, self).__init__(parent)
         self._application = application
